+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix error when reconnecting the client
 - Fix last modified timestamps using the remote node id
 - Add optional WebSocket heartbeat (on by default)
+- Improve logic for handling connected clients
 
 ## 0.0.1+1